Add hackish support for nodal forces
Nodal forces are handled through the addition of one-node elements
overlaid over existing nodes. However these elements are defined
one-by-one via commands that depend on the context (in particular,
the type of element currently active).

The implementation takes many shortcuts to identify the nodeset(s)
corresponding to the location of the force.